% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_countries}
\alias{fetch_countries}
\title{Fetch Countries}
\usage{
fetch_countries(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_countries hämtar alla länder finns att tillgå genom Börsdatas API
}
\description{
Fetch Countries
}
\examples{
key<-"<API KEY>"
fetch_countries(key=key)
}
