% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_instruments}
\alias{fetch_instruments}
\title{Fetch Instruments}
\usage{
fetch_instruments(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_instruments hämtar alla Instruments finns att tillgå genom Börsdatas API
}
\description{
Fetch Instruments
}
\examples{
key<-"<API KEY>"
fetch_instruments(key=key)
}
