% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_kpi_metadata}
\alias{fetch_kpi_metadata}
\title{Fetch KPI Metadata}
\usage{
fetch_kpi_metadata(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_kpi_metadata returnerar kpi metadata.
}
\description{
Fetch KPI Metadata
}
\examples{
key<-"<API KEY>"
fetch_kpi_metadata(key=key)
}
