% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_kpi_r12}
\alias{fetch_kpi_r12}
\title{Fetch KPI R12}
\usage{
fetch_kpi_r12(id = id, key = key)
}
\arguments{
\item{id}{Börsdata id}

\item{key}{Din börsdata API nyckel}
}
\value{
fetch_kpi_r12 returnerar de KPI nyckeltal som finns att tillgå genom
börsdatas api. Använder sig enbart av medelvärden, så det kan eventuellt vara av värde
att ändra om i funktionen för att hämta lägsta och högsta värden. Funktionen består av en
loop som går och hämtar ett nyckeltal i taget så det är ganska utmattade för API:n att hämta.
}
\description{
Fetch KPI R12
}
\examples{
key = "<API KEY>"
id = 750 # Exempel 750 för Evolution
fetch_kpi_r12(id=750,key=key)
}
