% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_quarter}
\alias{fetch_quarter}
\title{Fetch Quarter}
\usage{
fetch_quarter(id, key = key)
}
\arguments{
\item{id}{Börsdata id}

\item{key}{Din börsdata API nyckel}
}
\value{
fetch_quarter returnerar en data.frame med de 40 senaste kvartalsrapportna i sig. Returnerar 33 variabler.
}
\description{
Fetch Quarter
}
\examples{
key = "<API KEY>"
id = 750 # Exempel 750 för Evolution
fetch_quarter(id=750,key=key)
}
