% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_r12}
\alias{fetch_r12}
\title{Fetch R12}
\usage{
fetch_r12(id, key = key)
}
\arguments{
\item{id}{Börsdata id}

\item{key}{Din börsdata API nyckel}
}
\value{
fetch_r12 returnerar en data.frame med rullande 12 månders kvartalsrapporter i sig. Returnerar 33 variabler.
}
\description{
Fetch R12
}
\examples{
key = "<API KEY>"
id = 750 # Exempel 750 för Evolution
fetch_r12(id=750,key=key)
}
