% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_sectors}
\alias{fetch_sectors}
\title{Fetch Sectors}
\usage{
fetch_sectors(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_sectors hämtar alla Sektorer finns att tillgå genom Börsdatas API
}
\description{
Fetch Sectors
}
\examples{
key<-"<API KEY>"
fetch_sectors(key=key)
}
