% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_stockprice}
\alias{fetch_stockprice}
\title{Fetch Stockprice}
\usage{
fetch_stockprice(id, key = key)
}
\arguments{
\item{id}{Börsdata id}

\item{key}{Din börsdata API nyckel}
}
\value{
fetch_stockprice returnerar aktiekursen för bolaget per dag under de senaste 20 åren eller sedan dagen bolaget gick till börsen.
}
\description{
Fetch Stockprice
}
\examples{
key = "<API KEY>"
id = 750 # Exempel 750 för Evolution
fetch_stockprice(id=750,key=key)
}
