% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_stockprice_last}
\alias{fetch_stockprice_last}
\title{Fetch Stockprice Last}
\usage{
fetch_stockprice_last(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_stockprice_last returnerar den senaste aktiekursen för varje bolag.
}
\description{
Fetch Stockprice Last
}
\examples{
key<-"<API KEY>"
fetch_stockprice_last(key=key)
}
