% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_stocksplits}
\alias{fetch_stocksplits}
\title{Fetch Stock splits}
\usage{
fetch_stocksplits(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_stocksplits returnerar vilka bolag som har gjort en aktiesplit, storleken på splitten och åt vilket håll.
}
\description{
Fetch Stock splits
}
\examples{
key<-"<API KEY>"
fetch_stocksplits(key=key)
}
