% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_updated_instruments}
\alias{fetch_updated_instruments}
\title{Fetch Instruments Updated}
\usage{
fetch_updated_instruments(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_updated_instruments returnerar när olika bolag uppdaterade sina värden senast.
}
\description{
Fetch Instruments Updated
}
\examples{
key<-"<API KEY>"
fetch_updated_instruments(key=key)
}
