% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_updated_kpi_metadata}
\alias{fetch_updated_kpi_metadata}
\title{Fetch KPI Metadata Updated}
\usage{
fetch_updated_kpi_metadata(key = key)
}
\arguments{
\item{key}{Din börsdata API nyckel}
}
\value{
fetch_updated_kpi_metadata returnerar när metadatan har uppdaterats senast.
}
\description{
Fetch KPI Metadata Updated
}
\examples{
key<-"<API KEY>"
fetch_updated_kpi_metadata(key=key)
}
