% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borsdata.R
\name{fetch_year}
\alias{fetch_year}
\title{Fetch year}
\usage{
fetch_year(id, key = key)
}
\arguments{
\item{id}{Börsdata id}

\item{key}{Din börsdata API nyckel}
}
\value{
fetch_year returnerar de senaste 20 åren med årsredovisningar (eller så långt bak i tiden som det finns redovisningar).
}
\description{
Fetch year
}
\examples{
key = "<API KEY>"
id = 750 # Exempel 750 för Evolution
fetch_year(id=750,key=key)
}
